feat: start of workspace rework
