+++ src/components/modals/infoModal/styled.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const InfoModal = styled.div`
+  p {
+    margin-bottom: 8px;
+  }
+`;
