+++ src/components/modals/newPlanModal/NewPlanModal.tsx
@@ -1,0 +1,63 @@
+import { useState } from 'react';
+import { Button, Modal } from '@mantine/core';
+import { IconChevronsLeft } from '@tabler/icons';
+import * as styled from './styled';
+import { useWorkspaceStore } from '../../../store/workspaceStore';
+import { ROUTES } from '../../../App';
+import { useNavigate } from 'react-router-dom';
+
+const NewPlanModal = () => {
+  const navigate = useNavigate();
+  const resetWorkspace = useWorkspaceStore((state) => state.resetWorkspace);
+  const [opened, setOpened] = useState(false);
+
+  return (
+    <>
+      <Button
+        onClick={() => {
+          setOpened(!opened);
+        }}
+        leftIcon={<IconChevronsLeft />}
+        size='md'
+        radius='xl'
+      >
+        New floor plan
+      </Button>
+      <Modal
+        opened={opened}
+        onClose={() => setOpened(false)}
+        title='Discard current floorplan?'
+      >
+        <styled.NewPlanModal>
+          <p>
+            Are you sure you want to start over? You will lose all your current
+            progress.
+          </p>
+          <div>
+            <Button
+              onClick={() => {
+                setOpened(false);
+              }}
+              size='md'
+              radius='xl'
+            >
+              Cancel
+            </Button>
+            <Button
+              onClick={() => {
+                resetWorkspace();
+                navigate(ROUTES.HOME);
+              }}
+              size='md'
+              radius='xl'
+            >
+              Yes
+            </Button>
+          </div>
+        </styled.NewPlanModal>
+      </Modal>
+    </>
+  );
+};
+
+export default NewPlanModal;
