+++ src/components/modals/tallyModal/styled.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const TallyModal = styled.div``;
