+++ src/components/modals/touchWarningModal/styled.ts
