+++ src/pages/workspace/styled.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const WorkspaceSection = styled.section``;
+
+export const Topbar = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 16px;
+
+  > div {
+    display: flex;
+    align-items: center;
+    gap: 16px;
+  }
+`;
